% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_Matrix_Func.R
\name{all_k_suv}
\alias{all_k_suv}
\title{Title all survive prob}
\usage{
all_k_suv(k, n_start, n_stop, U_mat)
}
\arguments{
\item{k}{number of kin to survive}

\item{U_mat}{mortality matrix}

\item{s}{number of years kin survive (from birth)}
}
\value{
probability that "k" individuals survive for "s" years
}
\description{
Title all survive prob
}
