% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coa_PMF.R
\name{coa_PMF}
\alias{coa_PMF}
\title{Title Unconditional dist of Focal's cousins via older aunts}
\usage{
coa_PMF(y, s2, U_mat, F_mat, Q)
}
\arguments{
\item{y}{Scalar. age of Focal at present}

\item{s2}{Scalar. age of cousin at present (when Focal is y)}

\item{U_mat}{Matrix. probabilities of surviving from age class i to i+1 in the i-th entry of sub-diagonal}

\item{F_mat}{Matrix. fertility rate of age class i in the i-th entry of top-row}

\item{Q}{Scalar. Life-time maximum kin-number}
}
\value{
Vector. PMF for the number-distribution of cousins via older aunts of age s1 when Focal is y
}
\description{
Title Unconditional dist of Focal's cousins via older aunts
}
