% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/older_Aunts_PDF.R
\name{conditional_grans_age_OA}
\alias{conditional_grans_age_OA}
\title{Title}
\usage{
conditional_grans_age_OA(y, s1, b1, b2, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal at present}

\item{s1}{age of older aunt at present}

\item{b1}{age of mother when had Focal}

\item{b2}{age of gran when had mother}

\item{U_mat}{matrix of mortality probs}

\item{F_mat}{matrix of fertility rates}

\item{Q}{number of possible kin}
}
\value{
vector of probs of number of kin 0,1,2,..,Q
}
\description{
Title
}
