% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/older_Sisters_PDF.R
\name{conditional_mothers_age_OS}
\alias{conditional_mothers_age_OS}
\title{Title conditional distribution of older sisters, assuming we know mother's age at having Focal}
\usage{
conditional_mothers_age_OS(y, s1, b1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{(at present -- y in pdf document)}

\item{s1}{(at present -- s_1 in pdf document)}

\item{b1}{(age of mother when had focal, b_1 in pdf document)}

\item{U_mat}{(matrix of static survival probs -- subdiagonal)}

\item{F_mat}{(matrix of static fertility rates -- first row)}

\item{Q}{(the number of kin categories 0, 1, 2, ..., Q-1, i.e., max Q-1 kin over life)}
}
\value{
the pdf of older sisters of Focal age s_1, when focal is aged y, conditioned on mother being age b_1 at focal
}
\description{
Title conditional distribution of older sisters, assuming we know mother's age at having Focal
}
