% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/younger_Sisters_PDF.R
\name{conditional_mothers_age_YS}
\alias{conditional_mothers_age_YS}
\title{Title Conditional (on mother's age at Focal) pdf for Younger sisters}
\usage{
conditional_mothers_age_YS(y, s1, b1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal (at present)}

\item{s1}{age of younger sister (at present)}

\item{b1}{age of mother when having Focal}

\item{U_mat}{matrix of survival probabilities on sub-diagonal}

\item{F_mat}{matrix of age-specific fertilities on top-row}

\item{Q}{number of kin-number classes}
}
\value{
vector pdf of number / probs of younger sisters of age s1
}
\description{
Title Conditional (on mother's age at Focal) pdf for Younger sisters
}
