% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_convolution_Function.R
\name{convoluion_nth}
\alias{convoluion_nth}
\title{Title}
\usage{
convoluion_nth(n, list_ferts)
}
\arguments{
\item{n}{number of distributions}

\item{list_ferts}{list of the distributions}
}
\value{
Convolution over the distributions (Eq 3 in text)
}
\description{
Title
}
