% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/younger_cousin_Aunts_PDF.R
\name{CYA_dist_quick}
\alias{CYA_dist_quick}
\title{Title calculate the kin-number distribution of Focal's cousin descending from younger aunts}
\usage{
CYA_dist_quick(y, s2, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal}

\item{s2}{age of cousin from older aunt, when Focal is y}

\item{U_mat}{survival matrix}

\item{F_mat}{Fertilty matrix}

\item{Q}{number of possible kin}
}
\value{
vector in pmf form
}
\description{
Title calculate the kin-number distribution of Focal's cousin descending from younger aunts
}
