% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_pdfs_Func.R
\name{fert_dists}
\alias{fert_dists}
\title{Title}
\usage{
fert_dists(F_mat, Q)
}
\arguments{
\item{F_mat}{fertility matrix with age-specific rates on top-row}

\item{Q}{Maximum number of offspring}
}
\value{
Newborn probability distribution: Prob(number of newborns = 0,1,2,...,Q)
}
\description{
Title
}
