% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_fast_fourier_transform.R
\name{fourier_fft_nth}
\alias{fourier_fft_nth}
\title{Title}
\usage{
fourier_fft_nth(n, list_ferts)
}
\arguments{
\item{n}{number of distributions}

\item{list_ferts}{list of the distributions}
}
\value{
Convolution over the distributions (Eq 3 in text) by fft inverse
}
\description{
Title
}
