% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_Matrix_Func.R
\name{k_to_j}
\alias{k_to_j}
\title{Title prob that j out of k survive (Binomial expansion)}
\usage{
k_to_j(k, j, n_start, n_stop, U_mat)
}
\arguments{
\item{k}{number of kin}

\item{j}{number of kin survive (i.e., k - j die)}

\item{U_mat}{mortality matrix}

\item{s}{number of years}
}
\value{
probability that exactly "j out of k" individuals survive, over "s" years
}
\description{
Title prob that j out of k survive (Binomial expansion)
}
