% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/older_Sisters_PDF.R
\name{Matrix_func_age_OS}
\alias{Matrix_func_age_OS}
\title{Title Matrix implementation of the above "unconditional_mothers_age_OS"}
\usage{
Matrix_func_age_OS(y, s1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{(at present -- y in pdf document)}

\item{s1}{(at present -- s_1 in pdf document)}

\item{U_mat}{(matrix of static survival probs -- subdiagonal)}

\item{F_mat}{(matrix of static fertility rates -- first row)}

\item{Q}{(the number of kin categories 0, 1, 2, ..., Q-1, i.e., max Q-1 kin over life)}
}
\value{
vector with pmf of sisters aged s1 when Focal is y
}
\description{
Title Matrix implementation of the above "unconditional_mothers_age_OS"
}
