% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/younger_Aunts_PDF.R
\name{Matrix_func_age_YA_CALC}
\alias{Matrix_func_age_YA_CALC}
\title{Title matrix form the kin-number distribution of Focal's younger aunts}
\usage{
Matrix_func_age_YA_CALC(y, s1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal at present}

\item{s1}{age of younger aunt at present}

\item{U_mat}{U_mat matrix of mortality probs}

\item{F_mat}{matrix of fertility rates}

\item{Q}{kin number classes}
}
\value{
vector in pmf form
}
\description{
Title matrix form the kin-number distribution of Focal's younger aunts
}
