% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_mother_func.R
\name{mothers_age}
\alias{mothers_age}
\title{Title}
\usage{
mothers_age(U_mat, F_mat, x)
}
\arguments{
\item{U_mat}{survival matirx with age-specific probs of moving age x to x+1 on subdiagonal}

\item{F_mat}{fertility matirx with age-specific fertility rate on top-row}

\item{x}{possible age of mothering}
}
\value{
probability that a randomly selected newborn has a mother of age x
}
\description{
Title
}
