% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nys_PMF.R
\name{nys_PMF}
\alias{nys_PMF}
\title{Title Unconditional PDF for nieces through Focal's younger sisters}
\usage{
nys_PMF(y, s2, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal}

\item{s2}{age of younger niece when Focal is y}

\item{U_mat}{mort mat}

\item{F_mat}{fert mat}

\item{Q}{kin-numbers}
}
\value{
data frame with numbers and probs of kin aged s2
}
\description{
Title Unconditional PDF for nieces through Focal's younger sisters
}
