% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_PMF.R
\name{os_PMF_conditional}
\alias{os_PMF_conditional}
\title{Title Conditional PMF for Older sisters}
\usage{
os_PMF_conditional(y, s1, b1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal at present}

\item{s1}{age of Focal's older sister at present}

\item{b1}{(age of mother when had focal, b_1 in pdf document)}

\item{U_mat}{Matrix. probabilities of surviving from age class i to i+1 in the i-th entry of sub-diagonal}

\item{F_mat}{Matrix. fertility rate of age class i in the i-th entry of top-row}

\item{Q}{Scalar. Life-time maximum kin-number}
}
\value{
Vector. PMF for the number-distribution of older sisters of age s1 when Focal is y, conditional on mother's age b1
}
\description{
Title Conditional PMF for Older sisters
}
