% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fert_pdfs_Func.R
\name{pois_fert_pdf}
\alias{pois_fert_pdf}
\title{Title}
\usage{
pois_fert_pdf(r, k)
}
\arguments{
\item{r}{rate}

\item{k}{occurance}
}
\value{
Poisson distribution with parameters r,k
}
\description{
Title
}
