% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_matrix_Func.R
\name{Q_matrix}
\alias{Q_matrix}
\title{Title}
\usage{
Q_matrix(age, Q, F_mat)
}
\arguments{
\item{age}{age of individual at reproduction}

\item{Q}{upper bound for number of offspring}

\item{F_mat}{fertility matrix with age-specific rates on top-row}
}
\value{
The matrix F as in Eq 12 of manuscript. Projects a pmf of offrping,
}
\description{
Title
}
