% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_Matrix_Func.R
\name{U_kin_death}
\alias{U_kin_death}
\title{Title}
\usage{
U_kin_death(n_start, n_stop, Q, U_mat)
}
\arguments{
\item{Q}{number of kin-number classes (from 0 to Q-1 max)}

\item{U_mat}{mortality matrix}

\item{s}{years of survival}
}
\value{
Eqs 6,8,9 in manuscript. A matrix which projects the pdf of newborns for "s" years and
reshuffles them into kin-number classes conditional on death
}
\description{
Title
}
