% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/older_Aunts_PDF.R
\name{uncondition_grans_age_OA}
\alias{uncondition_grans_age_OA}
\title{Title}
\usage{
uncondition_grans_age_OA(y, s1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal at present}

\item{s1}{age of older aunt at present}

\item{U_mat}{U_mat matrix of mortality probs}

\item{F_mat}{matrix of fertility rates}

\item{Q}{kin number classes}
}
\value{
data frame: number of kin with corresponding probabilities of kin at age s1
}
\description{
Title
}
