% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestors_PDF.R
\name{unconditional_grandmother_pdf}
\alias{unconditional_grandmother_pdf}
\title{Title}
\usage{
unconditional_grandmother_pdf(y, s0, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal}

\item{s0}{age of gran when Focal is y}

\item{U_mat}{mortality matrix}

\item{F_mat}{fertility matrix}

\item{Q}{number of kin number classses}
}
\value{
a pdf in data frame output with numbers and resp probs
}
\description{
Title
}
