% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/older_cousin_Aunts_PDF.R
\name{unconditional_grans_age_COA}
\alias{unconditional_grans_age_COA}
\title{Title}
\usage{
unconditional_grans_age_COA(y, s2, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal}

\item{s2}{age of cousin from older aunt, when Focal is y}

\item{U_mat}{survival matrix}

\item{F_mat}{Fertilty matrix}

\item{Q}{number of possible kin}
}
\value{
data frame: number of kin with corresponding probabilities, for kin at age s2
}
\description{
Title
}
