% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/younger_cousin_Aunts_PDF.R
\name{unconditional_grans_age_CYA}
\alias{unconditional_grans_age_CYA}
\title{Title calculate the kin-number distribution of Focal's cousin descending from younger aunts}
\usage{
unconditional_grans_age_CYA(y, s2, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal}

\item{s2}{age of Focal's cousin from younger aunt, when Focal is y}

\item{U_mat}{matrix with survival probs}

\item{F_mat}{matrix of fertility rates}

\item{Q}{number of kin classes (0,1,2,...,Q)}
}
\value{
data frame: number of kin and probability for kin of age s2
}
\description{
Title calculate the kin-number distribution of Focal's cousin descending from younger aunts
}
