% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/younger_Aunts_PDF.R
\name{unconditional_grans_age_YA}
\alias{unconditional_grans_age_YA}
\title{Title unconditional dist of Focal's younger aunts:}
\usage{
unconditional_grans_age_YA(y, s1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal}

\item{s1}{age of younger aunt (both at present age FOcal)}

\item{U_mat}{mort mat}

\item{F_mat}{fert mat}

\item{Q}{kin-numbers}
}
\value{
data frame: numbers and probs of kin at age s1
}
\description{
Title unconditional dist of Focal's younger aunts:
}
