% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestors_PDF.R
\name{unconditional_mother_pdf}
\alias{unconditional_mother_pdf}
\title{Title}
\usage{
unconditional_mother_pdf(y, s0, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal}

\item{s0}{age of mother when Focal is y}

\item{U_mat}{matrix of mortality rates}

\item{F_mat}{matrix of fertility rates}

\item{Q}{number of kin classes 0,...,Q-1}
}
\value{
an age-specific pdf (data frame) with number = 0, 1, ..., Q-1 and resp. probs of kin
}
\description{
Title
}
