% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/older_niece_Sisters_PDF.R
\name{unconditional_mothers_age_NOS}
\alias{unconditional_mothers_age_NOS}
\title{Title Unconditional PDF for nieces through Focal's older sisters}
\usage{
unconditional_mothers_age_NOS(y, s2, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal}

\item{s2}{age of older niece when Focal is y}

\item{U_mat}{mort mat}

\item{F_mat}{fert mat}

\item{Q}{kin-numbers}
}
\value{
data frame with numbers and probs of kin aged s2
}
\description{
Title Unconditional PDF for nieces through Focal's older sisters
}
