% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/older_Sisters_PDF.R
\name{unconditional_mothers_age_OS}
\alias{unconditional_mothers_age_OS}
\title{Title unconditional distribution of older sisters, weighted over all ages when mother had Focal}
\usage{
unconditional_mothers_age_OS(y, s1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{(at present -- y in pdf document)}

\item{s1}{(at present -- s_1 in pdf document)}

\item{U_mat}{(matrix of static survival probs -- subdiagonal)}

\item{F_mat}{(matrix of static fertility rates -- first row)}

\item{Q}{(the number of kin categories 0, 1, 2, ..., Q-1, i.e., max Q-1 kin over life)}
}
\value{
data frame with columns: numbers 0,..,Q-1, probs of numbers,
}
\description{
Title unconditional distribution of older sisters, weighted over all ages when mother had Focal
}
