% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/younger_Sisters_PDF.R
\name{unconditional_mothers_age_YS}
\alias{unconditional_mothers_age_YS}
\title{Title}
\usage{
unconditional_mothers_age_YS(y, s1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal (at present)}

\item{s1}{age of younger sister (at present)}

\item{U_mat}{matrix of survival probabilities on sub-diagonal}

\item{F_mat}{matrix of age-specific fertilities on top-row}

\item{Q}{number of kin-number classes
\subsection{Note that we sum over probable ages of mother in this case}{
}}
}
\value{
data.fram of pdf of number / probs of younger sisters of age s1
}
\description{
Title
}
