% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_PMF.R
\name{ya_PMF}
\alias{ya_PMF}
\title{Title Unconditional dist of Focal's younger aunts}
\usage{
ya_PMF(y, s1, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal at present}

\item{s1}{age of younger aunt at present}

\item{U_mat}{Matrix. probabilities of surviving from age class i to i+1 in the i-th entry of sub-diagonal}

\item{F_mat}{Matrix. fertility rate of age class i in the i-th entry of top-row}

\item{Q}{Scalar. Life-time maximum kin-number}
}
\value{
Vector. PMF for the number-distribution of younger aunts of age s1 when Focal is y
}
\description{
Title Unconditional dist of Focal's younger aunts
}
