% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ya_PMF.R
\name{ya_PMF_conditional}
\alias{ya_PMF_conditional}
\title{Title Conditional dist of Focal's younger aunts: conditioned on mother age b1 and gran age b2}
\usage{
ya_PMF_conditional(y, s1, b1, b2, U_mat, F_mat, Q)
}
\arguments{
\item{y}{age of Focal at present}

\item{s1}{age of younger aunt at present}

\item{b1}{age of mother when had Focal}

\item{b2}{age of gran when had mother}

\item{U_mat}{Matrix. probabilities of surviving from age class i to i+1 in the i-th entry of sub-diagonal}

\item{F_mat}{Matrix. fertility rate of age class i in the i-th entry of top-row}

\item{Q}{Scalar. Life-time maximum kin-number}
}
\value{
Vector. PMF for the number-distribution of younger aunts of age s1 when Focal is y, conditional on both mother's age b1 and gran's age b2
}
\description{
Title Conditional dist of Focal's younger aunts: conditioned on mother age b1 and gran age b2
}
